trash functionality for collections
When a collection is deleted, it goes to trash where it shows
up among deleted items. Subcollections of a deleted collection
go to trash as well.
Subcollections of a deleted collection do not show up in trash -
only top-most collection does.
When a collection is deleted or restored, its children (items +
subcollections) are deleted or restored in case if they were
moved to trash at the same time as the collection.
On select of a collection in the trash, the context pane
displays a message with the number of subcollections and items
that were moved to trash at the same time with the specified collection.