aria properties to improve accessibility
Added aria roles and labels to make context pane
recognizable for wooshy, voiceover and jaws.
Pointed context-pane fields to their labels so that
screen readers read the label on input focus.
Minor bug fix to not increment lastFocusedField tab index
if the focus does not change (that prevents focus from being
stuck on the last field).
Added aria label to reader' browser so that screen reader can read
the title when tabs are switched.
Fixes: #2736
Partly fixes:  #3250