Fix recognizeDocument test failures
Wait for all progress-queue rows to be done processing before moving on
to the next test. Without this, preview rendering or other operations
can cause test failures by delaying the `ZoteroPane.selectItems()` call
for the new parent item in `_processItem()` until the middle of a
following test (due to the await for file renaming [1]). If it's delayed
until after the next attachment has been created, the previous parent
item will be selected after the new attachment and `recognizeSelected()`
in the test won't work. This is most pronounced with the reader, but it
was apparently happening previously due to something else, hence the
explicit item selection (now removed) in one test.

[1] https://github.com/zotero/zotero/blob/21e50add60ee1533cb7e528c7edc08f2f1ccddb1/chrome/content/zotero/xpcom/recognizeDocument.js#L289-L301