Fix attachment column sorting not always working.
This commit reduces the performance of the tree view.

Sorting by the `hasAttachment` column depends on having the `bestAttachmentState`
determined for every item in the current view. Normally, this is not available
for rows that are not currently visible, for performance reasons. There was a
piece of logic that triggered `bestAttachmentState` logic for every
item in the current view when the user initiated sorting by the `hasAttachment`
column. Once this logic was executed, the tree set `_cachedBestAttachmentStates`
to true and from that point onwards assumed that the `bestAttachmentState` was
known for every item in the view.

However, to the best of my knowledge, this assumption is not always correct.
Firstly, when the user navigates to another collection or library, the
assumption may no longer hold true. Additionally, this method will not work
after, for example, Zotero is restarted, in which case the sorting may have been
configured to the `hasAttachment` column before the restart, but the
`_cachedBestAttachmentStates` logic will not be triggered.

In this commit, I have moved that logic to the `sort` method. However, for a very
large library, when many items with attachments are present in the current view,
this might introduce some lag on less powerful hardware. This issue could
perhaps be addressed by caching `bestAttachment` at the database layer?