New approach to sticking/unsticking
I think this is overall better and more robust. It fixes a bunch of
weird issues when switching between >100%-height titles and regular
titles, it scrolls to the correct place every time, and the performance
impact should be minimal because we were already using a scroll
listener, not just the IntersectionObserver.

The caveat is that we need to constrain titles to 100% height at most.
You can still scroll within the editable-text! I'd like titles to be
able to keep growing and just scroll along with the pane contents, but
I can't find a sane way to make that work with scroll padding
calculation.