reworked locatorNode logic
Shortcuts, such as "p. X", or ":X", will always apply X locator
value to the bubble right before the input. The only exception is when a
new bubble was just added from the reference panel - in this case, the
shortcuts will stick to the new bubble. This effect resets on the blur
of the input field.