initial setup of annotations showing in itemTree
- Annotations are displayed in itemTree under their file attachments
on a third level. The annotation spans the entire row.
- Search matches the actual attachment instead of it's parent file.
- Can create child notes from annotations of the same item or
standalone notes from annotations across different items from
the context menu.

Need to:
- add something to be displayed in the itemPane when annotation is selected.
Naturally, the preview of the annotation and ... what else?
- tweak the title to be displayed better. We need to show as much of
the comment as possible but also leave at least a bit of room for the
annotation text.
Ink annotations should have no text and no comments so it should be just
"Ink annotation" but we could show some tags.