qf: edits to aria item-description
- include a sentence that contains which section the item belongs to
(e.g. selected/cited/my library) before the keypresses hint of the item
description. The section is updated whenever an item is focused.
- if no item from the ref panel is focused, the section is empty and no
item-description is announced. Since the panel is a popup, NVDA and JAWS sometimes
begin reading out all items from the reference panel as soon as it appears.
In that case, keypress hints do not make sense, so it's best to just
skip the aria-description.

Fixes: #3764