on selection change on right-click in virt table
- on right key down in virtualized table, record the
receiving table row as "contextMenuRow"
- contextMenuRow is cleared when the main popup
disappears
- when a selected row is fetched in itemTree/collectionTree
or ZoteroPane, return the contextMenuRow if it exists.
It ensures that the selected action will run on the
right-clicked row and not the row currently selected
- some actions run only after the confirmation from
a dialog. The contextMenuRow wants to be cleared
when the menupopup is hidden. To make sure it remains
until the dialgo is closed, create a promise that
is awaited in popuphidden handler, which is resolved
when the dialog is confirmed/cancelled. ZoteroPane
has a wrapper function through which all dialogs
are handled with the promises
- minor tweaks to run zotero pane functions when the
selection is not empty OR when there is a context menu
row