qf: fix locator node gliitches
- Fixed a bug where if you add a bubble, immediately
delete it and then type "p. X", the page number would not
be added anywhere because the locator was still tied
to the removed bubble.
- On click between bubbles, the old input is removed directly
and not via 'blur' listener. To not have the locator
node being stuck to a previously added bubble,
it is now cleared before focusing the new input.