fx115: fixes to qf reference panel positioning
- Added flip="none" to reference panel to stop it from
opening above the dialog or sliding through it.
- On linux and windows, when the window is dragged, the reference panel
remains where it is and does not follow the dialog. To prevent the panel
from "breaking away" from the main window, hide it when the drag starts
and reopen when dragging is done
- On windows, the mouse often "outruns" the window when it is being dragged.
If that happens, the dialog stops moving and it can miss the
'mouseup' event. In that case, the panel won't reopen and the dialog will
start being dragged on next 'mouseover' without clicking until the next
'mousedown' event. This has been an issue before fx115 as well but if
panel opening/closing is a part of the window drag logic, it's best
to have it addressed. There should be a better fix to make dialog dragging
work better on windows overall but this is a start.