Preferences: Restore subpane Back button
It was technically still there, just offscreen. The wrapper element
was display: block, and some block layout quirk caused the button to
grow to fill the wrapper, which then put it offscreen when rotated.

This removes the wrapper and replaces it with a margin. Also fixes an
incorrect margin in RTL locales.