only reset scrollToCell state when needed
We really need to reset scrollToCell only when the
focused tag has not changed. In all other cases, scrollToCell
prop should always be different on each re-render.

An example of this is: focus the first tag, scroll
all the way down, press arrowLeft to scroll to the top,
and this is when we need to reset scrollToCell