vpat 19: aria labels for filter inputs
The filters only have a placeholder, so once there
is some text, it goes away and nothing else gets
announced, so we need a label. However, the note
from the VPAT review indicates that the label needs
to be VISIBLE per 3.3.2 success criteria:
https://www.w3.org/WAI/WCAG21/Understanding/labels-or-instructions.

This is something worth discussing, since 3.3.2 seems
mainly concerned with inputs for form fields, while
these are search fields which is somewhat different.
So I am not sure if the visibility requirement
necessarily applies here. After all, https://www.w3.org/WAI/
that lists the accessibility criteria has a search
field with the label not being visible ...